# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
  /auth/register:
    post:
      operationId: registerAccount
      tags: [auth]
      summary: Register a user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "400":
          description: Invalid registration data
        "409":
          description: Email already registered
  /auth/login:
    post:
      operationId: loginAccount
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getCurrentAccount
      tags: [auth]
      summary: Get the current signed-in account
      responses:
        "200":
          description: Current account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Authentication required
...
  /analytics/events:
    post:
      operationId: trackAnalyticsEvent
      tags: [analytics]
      summary: Track a first-party analytics event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyticsEventInput"
      responses:
        "204":
          description: Event accepted
        "400":
          description: Invalid event
...
  /storage/uploads/request-url:
    post:
      operationId: requestUploadUrl
      tags: [storage]
      summary: Request a presigned article image upload URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid upload metadata
        "401":
          description: Admin access required
        "500":
          description: Failed to generate upload URL
...
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve an uploaded object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stored object
          content:
            "*/*":
              schema:
                type: string
                format: binary
        "404":
          description: Object not found
components:
  schemas:
    Account:
      type: object
      required: [id, name, email, role]
      properties:
        id: { type: number }
        name: { type: string }
        email: { type: string, format: email }
        role: { type: string }
    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, minLength: 2, maxLength: 80 }
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    AnalyticsEventInput:
      type: object
      required: [eventType, path]
      properties:
        eventType:
          type: string
          enum: [page_view, article_view, search, newsletter_signup, share]
        path: { type: string, maxLength: 500 }
        articleId: { type: integer, nullable: true }
        referrer: { type: string, maxLength: 1000 }
        sessionId: { type: string, maxLength: 128 }
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: number
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
        objectPath:
          type: string
        metadata:
          $ref: "#/components/schemas/UploadUrlRequest"